+++ ElvUI_EltreumUI/Modules/Skins/Borders.lua
@@ -147,6 +147,20 @@ function ElvUI_EltreumUI:Borders()
 				targettargetborder:SetBackdropBorderColor(classcolor.r, classcolor.g, classcolor.b, 1)
 				targettargetborder:SetFrameStrata("LOW")
 			end
+
+			if E.db.ElvUI_EltreumUI.borders.petborder and E.db.unitframe.units.pet.enable then
+				local petborder = CreateFrame("Frame", "EltruismPetBorder", _G.ElvUF_Pet_HealthBar, BackdropTemplateMixin and "BackdropTemplate")
+				local petsizex, petsizey = _G["ElvUF_Pet_HealthBar"]:GetSize()
+				petborder:SetSize(petsizex, petsizey)
+				petborder:SetPoint("TOPRIGHT", _G.ElvUF_Pet_HealthBar,"TOPRIGHT", 17, 16)
+				petborder:SetPoint("BOTTOMLEFT", _G.ElvUF_Pet_HealthBar,"BOTTOMLEFT", -17, -16)
+				petborder:SetBackdrop({
+					edgeFile = bordertexture,
+					edgeSize = playertargetsize,
+				})
+				petborder:SetBackdropBorderColor(classcolor.r, classcolor.g, classcolor.b, 1)
+				petborder:SetFrameStrata("LOW")
+			end
 		end
 
 		--elvui action bars
@@ -292,9 +306,54 @@ function ElvUI_EltreumUI:Borders()
 			powerbarborder:SetBackdropBorderColor(classcolor.r, classcolor.g, classcolor.b, 1)
 			powerbarborder:SetFrameStrata("HIGH")
 		end
+
+		-- minimap
+		if E.private["general"]["minimap"]["enable"] ~= false and E.db.ElvUI_EltreumUI.borders.minimapborder then
+			local MinimapBorder = CreateFrame("Frame", "EltruismMiniMapBorderFrame", _G["Minimap"], BackdropTemplateMixin and "BackdropTemplate")
+			local Minimapsizex, Minimapsizey = _G["Minimap"]:GetSize()
+			MinimapBorder:SetSize(Minimapsizex, Minimapsizey)
+			MinimapBorder:SetParent(_G["Minimap"])
+			if not (self.minimaphasBorder) then
+				MinimapBorder:SetPoint("TOPRIGHT", _G["Minimap"] ,"TOPRIGHT", 13, 13)
+				MinimapBorder:SetPoint("BOTTOMLEFT", _G["MinimapPanel"] ,"BOTTOMLEFT", -13, -13)
+				MinimapBorder:SetBackdrop({
+					edgeFile = bordertexture,
+					--edgeSize = E.db.ElvUI_EltreumUI.borders.baredgesize, --13
+					edgeSize = 15,
+				})
+				MinimapBorder:SetBackdropBorderColor(classcolor.r, classcolor.g, classcolor.b, 1)
+				MinimapBorder:SetFrameStrata("HIGH")
+				if E.db.datatexts.panels.MinimapPanel.backdrop == false or E.db.datatexts.panels.MinimapPanel.enable == false then
+					MinimapBorder:SetPoint("BOTTOMLEFT", _G["Minimap"] ,"BOTTOMLEFT", -13, -13)
+				end
+				if IsAddOnLoaded("ElvUI_SLE") and E.private["sle"]["minimap"]["rectangle"] == true then --Shadow and Light Rectangle Minimap
+					MinimapBorder:SetPoint("TOPRIGHT", _G["Minimap"].backdrop ,"TOPRIGHT", 12, 11)
+					MinimapBorder:SetPoint("BOTTOMLEFT", _G["MinimapPanel"] ,"BOTTOMLEFT", -13, -13)
+					if E.db.datatexts.panels.MinimapPanel.backdrop == false or E.db.datatexts.panels.MinimapPanel.enable == false then
+						MinimapBorder:SetPoint("BOTTOMLEFT", _G["Minimap"].backdrop ,"BOTTOMLEFT", -11, -11)
+					end
+					local updatelocationpos = CreateFrame("Frame")
+					updatelocationpos:RegisterEvent("ZONE_CHANGED")
+					updatelocationpos:RegisterEvent("ZONE_CHANGED_INDOORS")
+					updatelocationpos:RegisterEvent("ZONE_CHANGED_NEW_AREA")
+					updatelocationpos:RegisterEvent("PLAYER_ENTERING_WORLD")
+					updatelocationpos:RegisterEvent("MINIMAP_UPDATE_ZOOM")
+					updatelocationpos:SetScript("OnEvent", function()
+						_G.Minimap.location:ClearAllPoints()
+						_G.Minimap.location:SetPoint('TOP', _G.Minimap, 'TOP', 0, -15)
+					end)
+
+				elseif IsAddOnLoaded("ElvUI_WindTools") and E.db["WT"]["maps"]["rectangleMinimap"]["enable"] == true then --Windtools rectangle minimap
+					MinimapBorder:SetPoint("TOPRIGHT", _G["MinimapBackdrop"] ,"TOPRIGHT", 12, 11)
+					MinimapBorder:SetPoint("BOTTOMLEFT", _G["MinimapBackdrop"] ,"BOTTOMLEFT", -12, -12)
+				end
+				MinimapBorder:Show()
+				self.minimaphasBorder = true
+			end
+		end
+
 	end
 end
-
 
 function ElvUI_EltreumUI:BordersTargetChanged()
 	if not IsAddOnLoaded("ElvUI_EltreumUI") then