+++ ElvUI_EltreumUI/Profiles/Capping.lua
@@ -38,5 +38,8 @@ function ElvUI_EltreumUI:GetCappingProfile()
 				},
 			}
 		end
+
+
+		ElvUI_EltreumUI:Print(L["Capping profile has been set."])
 	end
 end